🎨 Moved mineCheck into a separate file
